Have sequence sorting actually use the value and not the object address (doh)

+++ Classes/Stop+Convenience.m
@@ -12,11 +12,11 @@
 
 - (NSComparisonResult)sortBySequence:(Stop *)otherStop
 {
-	if(self.stopSequence == otherStop.stopSequence)
+	if(self.stopSequence.intValue == otherStop.stopSequence.intValue)
     {
         return NSOrderedSame;
     }
-    else if(self.stopSequence < otherStop.stopSequence)
+    else if(self.stopSequence.intValue < otherStop.stopSequence.intValue)
     {
         return NSOrderedAscending;
     }
